+++ lib/ORM/Relation/BelongsTo.php
@@ -62,7 +62,7 @@ class BelongsTo extends Relation
         $callback = $options['callback'];
         $immediate = $options['immediate'];
 
-        $select = new Select($this->compiler, $this->model->getTable());
+        $select = new Select($this->model, $this->compiler);
 
         $select->where($pk)->in($ids);
 
@@ -89,12 +89,20 @@ class BelongsTo extends Relation
     }
 
     /**
+     * Build query
+     * 
+     * @return  Select
+     */
+    protected function buildQuery()
+    {
+        return $this->query->where($this->model->getPrimaryKey())->is($this->owner->{$this->getForeignKey()});
+    }
+
+    /**
      * @return  Model
      */
     public function getResult()
     {
-        $this->query->where($this->model->getPrimaryKey())->is($this->owner->{$this->getForeignKey()});
-
         return $this->query()
                 ->fetchClass(get_class($this->model), array($this->isReadOnly, $this->connection))
                 ->first();