+++ lib/ORM/Relation/HasOne.php
@@ -38,8 +38,6 @@ class HasOne extends Relation
      */
     public function getResult()
     {
-        $this->query->where($this->getForeignKey())->is($this->owner->{$this->owner->getPrimaryKey()});
-
         return $this->query()
                 ->fetchClass(get_class($this->model), array($this->isReadOnly, $this->connection))
                 ->first();