+++ lib/ORM/Update.php
@@ -21,29 +21,28 @@
 namespace Opis\Database\ORM;
 
 use Opis\Database\Connection;
-use Opis\Database\SQL\Compiler;
-use Opis\Database\SQL\WhereClause;
 use Opis\Database\SQL\UpdateStatement;
 
 class Update extends UpdateStatement
 {
+    /** @var    Select */
+    protected $select;
+
     /** @var    Connection */
     protected $connection;
 
     /**
      * Constructor
      *
-     * @param   Connection          $connection
-     * @param   Compiler            $compiler
-     * @param   string              $from
-     * @param   array               $joins
-     * @param   WhereClause|null    $clause     (optional)
+     * @param   Slect       $select
+     * @param   Connection  $connection
      */
-    public function __construct(Connection $connection, Compiler $compiler, $from, $joins, WhereClause $clause = null)
+    public function __construct(Select $select, Connection $connection)
     {
-        parent::__construct($compiler, $from, $clause);
+        parent::__construct($select->getCompiler(), $select->getTables(), $select->getWhereClause());
+        $this->joins = $select->getJoinClauses();
         $this->connection = $connection;
-        $this->joins = $joins;
+        $this->select = $select;
     }
 
     /**
@@ -51,8 +50,17 @@ class Update extends UpdateStatement
      *
      * @return  int
      */
-    public function update(array $columns)
+    public function update(array $columns, $softDelete = false)
     {
+        if ($this->select->supportsSoftDeletes()) {
+
+            if (!$this->select->isSetInlcudeSoftDeletes()) {
+                $this->where('deleted_at')->isNull();
+            } elseif ($this->select->isSetOnlySoftDeleted()) {
+                $this->where('deleted_at')->notNull();
+            }
+        }
+
         parent::set($columns);
         return $this->connection->count((string) $this, $this->compiler->getParams());
     }