+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 CHANGELOG
 -----------
+
+* Added support for soft deletes
+* `Opis\Database\Model::softDelete`, `Opis\Database\Model::supportsSoftDeletes`, `Opis\Database\ORM\Query::softDelete`,
+`Opis\Database\ORM\Query::restore`, `Opis\Database\ORM\BaseQuery::withSoftDeleted`, 
+`Opis\Database\ORM\BaseQuery::onlySoftDeleted` methods were added.
+* `Opis\Database\Model::destroy` and `Opis\Database\Model::softDestroy` methods were added
+* The following methods were added to `Opis\Database\ORM\Relation` class: `column`, `count`, `sum`, `avg`, `min`, `max`,
+`update`, `restore`, `delete` and `softDelete`
+* Now you can count, update, delete, soft delete and restore related models
+
 ### v3.4.2, 2016.02.23
 
 * Fixed a bug in `Opis\Database\Model::__set` method